# repo/backend/openapi.yaml
openapi: 3.0.3
info:
  title: MeridianCare Offline API
  version: 1.0.0
servers:
  - url: http://localhost:13000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        msg:
          type: string
      required: [code, msg]
    AuthLoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            username:
              type: string
            role:
              type: string
            clinic_id:
              type: string
              format: uuid
              nullable: true
      required: [token, role]
    Patient:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        allergies:
          type: array
          items:
            type: object
        contraindications:
          type: array
          items:
            type: object
        clinic_id:
          type: string
          format: uuid
          nullable: true
      required: [id, name]
    PatientCreateRequest:
      type: object
      properties:
        name:
          type: string
        ssn:
          type: string
        allergies:
          type: array
          items:
            type: object
        contraindications:
          type: array
          items:
            type: object
      required: [name]
    Prescription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        encounter_id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        drug_name:
          type: string
        dose:
          type: string
        route:
          type: string
        quantity:
          type: integer
        state:
          type: string
        version:
          type: integer
      required: [id, patient_id, drug_name, state, version]
    PrescriptionCreateRequest:
      type: object
      properties:
        encounterId:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        drugName:
          type: string
        dose:
          type: string
        route:
          type: string
        quantity:
          type: integer
        instructions:
          type: string
        overrideReason:
          type: string
        reauthPassword:
          type: string
      required: [encounterId, patientId, drugName, dose, route, quantity, instructions]
    PharmacyActionRequest:
      type: object
      properties:
        action:
          type: string
          enum: [approve, dispense, void]
        expectedVersion:
          type: integer
        inventoryItemId:
          type: string
          format: uuid
        dispenseQuantity:
          type: integer
        reason:
          type: string
      required: [action]
    CredentialingImportRequest:
      type: object
      properties:
        rows:
          type: array
          items:
            type: object
        mapping:
          type: object
          additionalProperties:
            type: string
      required: [rows]
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
  /api/auth/login:
    post:
      summary: Local login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Auth success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/me:
    get:
      summary: Current authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/patients:
    post:
      summary: Create patient
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientCreateRequest'
      responses:
        '201':
          description: Created patient
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/prescriptions:
    post:
      summary: Submit prescription with hard-stop override policy
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrescriptionCreateRequest'
      responses:
        '201':
          description: Created prescription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prescription'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/pharmacy/queue:
    get:
      summary: Pharmacist review queue with inventory availability
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Queue rows
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/pharmacy/{id}/action:
    post:
      summary: Approve/dispense/void transitions
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PharmacyActionRequest'
      responses:
        '200':
          description: Updated prescription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prescription'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/credentialing/import:
    post:
      summary: Batch import with row-level validation errors
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialingImportRequest'
      responses:
        '200':
          description: Import result
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'